backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.1
websockets>=12.0
google-genai>=0.2.2
google-cloud-firestore>=2.16.0
python-dotenv>=1.0.1
pydantic>=2.7.0
google-auth>=2.29.0
passlib[bcrypt]
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.9
requests>=2.31.0
beautifulsoup4>=4.12.3
